Configure API PATCH AND POST Requests For Class Routes
Previously the Patch and Post request were not configured. This commit
will handle the requests coming into those two methods.

This commit also refactors some code to keep it dry.  The try catch
block was moved to it's own function and the appropriate callbacks and
parameters were set to allow it to be used universally across the Class
routes.